feat: implement useLocalStorage hook
